Release using PEX and nfpm instead of dh-virtualenv
This is significantly simpler, more flexible and less weird than
using dh-virtualenv.

PEX builds a single executable with all Python dependencies and nfpm
wraps that up in a .deb with the config template and systemd service
files. The `build` script automates the build process.

The debian build system cruft is gone.